Extract Review tabs into dedicated views
Split the inline tab bodies out of Review/ContentView.cs so each tab is its
own view with clear separation of concerns:

- Add Review/Tabs/PlanTabView.cs (read-only plan markdown tab).
- Add Review/Tabs/RecommendationsTabView.cs (Implement button + row list);
  RecommendationRowView moves in alongside its only consumer.
- Add Review/ReviewActionsBarView.cs (project review-actions bar; not a tab,
  so it lives directly under Review/).
- ContentView.BuildContent now composes these views while keeping ownership of
  the tab assembly, navigation, static helpers, and overlay sheets.

Shared Views/Tabs/* views are left untouched (Drafts still depends on them).
No behavior change; build clean and ContentViewTests green.
